Added logging and refactored backup process

+++ Izio.Umbraco.SourceControlLibraries/Backup.cs
@@ -11,8 +11,7 @@ namespace Izio.Umbraco.SourceControlLibraries
 {
     public class Backup : ApplicationEventHandler
     {
-        protected override void ApplicationStarted(UmbracoApplicationBase umbracoApplication,
-            ApplicationContext applicationContext)
+        protected override void ApplicationStarted(UmbracoApplicationBase umbracoApplication, ApplicationContext applicationContext)
         {
             var installedPackagesConfig = XDocument.Load(HostingEnvironment.MapPath("~/App_Data/packages/installed/installedPackages.config"));
             var installedLibraries = installedPackagesConfig.Descendants("file").Where(f => f.Value.ToLower().EndsWith("dll")).Select(f => f.Value.Substring(5)).Distinct();
@@ -21,16 +20,31 @@ namespace Izio.Umbraco.SourceControlLibraries
             //create backup folder if it doesn't exist
             if (Directory.Exists(HostingEnvironment.MapPath("~/" + configuration.Folder)) == false)
             {
-                Directory.CreateDirectory(HostingEnvironment.MapPath("~/" + configuration.Folder));
+                try
+                {
+                    LogHelper.Info<Backup>("Creating backup folder");
+
+                    Directory.CreateDirectory(HostingEnvironment.MapPath("~/" + configuration.Folder));
+                }
+                catch (Exception ex)
+                {
+                    LogHelper.Error<Backup>("Failed to create backup folder", ex);
+
+                    throw;
+                }
             }
 
             //delete existing backups that are no longer required and remove from configuration
             foreach (var library in Directory.GetFiles(HostingEnvironment.MapPath("~/" + configuration.Folder)))
             {
+                LogHelper.Info<Backup>("Deleting expired backups");
+
                 var fileName = Path.GetFileName(library);
 
                 if (installedLibraries.Contains(fileName) == false)
                 {
+                    LogHelper.Info<Backup>(string.Format("Removing {0}", fileName));
+
                     File.Delete(library);
 
                     configuration.Libraries.Remove(fileName);
@@ -40,19 +54,38 @@ namespace Izio.Umbraco.SourceControlLibraries
             //iterate over all libraries
             foreach (var library in installedLibraries)
             {
+                //check if library exists in bin folder
                 if (File.Exists(HostingEnvironment.MapPath("~/bin/" + library)))
                 {
-                    //copy library to backup location
-                    File.Copy(HostingEnvironment.MapPath("~/bin/" + library), HostingEnvironment.MapPath("~/" + configuration.Folder + "/" + library), true);
+                    //copy library to backup folder if it doesn't exist
+                    if (File.Exists(HostingEnvironment.MapPath("~/" + configuration.Folder + "/" + library)) == false)
+                    {
+                        LogHelper.Info<Backup>(string.Format("Backing up {0}", library));
+
+                        //copy library to backup location
+                        File.Copy(HostingEnvironment.MapPath("~/bin/" + library), HostingEnvironment.MapPath("~/" + configuration.Folder + "/" + library), true);
+                    }
+
+                    //add to configuration list if it doesn't exist
+                    if (configuration.Libraries.Contains(library) == false)
+                    {
+                        LogHelper.Info<Backup>(string.Format("Adding {0} to configuration", library));
+
+                        configuration.Libraries.Add(library);
+                    }
                 }
                 else
                 {
+                    LogHelper.Info<Backup>(string.Format("Restoring {0}", library));
+
                     //restore library from backup location
                     File.Copy(HostingEnvironment.MapPath("~/" + configuration.Folder + "/" + library), HostingEnvironment.MapPath("~/bin/" + library), true);
 
                     //execute packages if specified
                     if (configuration.ExecutePackages)
                     {
+                        LogHelper.Info<Backup>(string.Format("Restoring {0}", library));
+
                         //load plugin assembly
                         var plugin = System.Reflection.Assembly.LoadFile(HostingEnvironment.MapPath("~/bin/" + library));
 
@@ -64,6 +97,8 @@ namespace Izio.Umbraco.SourceControlLibraries
                         {
                             try
                             {
+                                LogHelper.Info<Backup>(string.Format("Executing package actions {0}", instance.Alias()));
+
                                 instance.Execute(instance.Alias(), instance.SampleXml());
                             }
                             catch (Exception ex)
@@ -72,11 +107,6 @@ namespace Izio.Umbraco.SourceControlLibraries
                             }
                         }
                     }
-                }
-
-                if (configuration.Libraries.Contains(library) == false)
-                {
-                    configuration.Libraries.Add(library);
                 }
             }
